feat: implement support for UBX-MON-COMMS
This message allows monitoring the board's internal comms subsystems.
I have implemented it so that we can debug the issues related to missing
RTCM messages on the base station output.

Indeed, the problem seemed to be the UART1 being overloaded (having
a rather low bitrate), which impacts other channels like USB. Turning
messages off on it solved our issue.