feat: use long pulse periods to disambiguate chrony info when RTK goes bad
The issue this suite of commits is solving is that the RTK
engine may wait "up to one second" for correction info before giving
up and calculating a solution, when e.g. the RTCM messages only
partially arrive. This has already confused chrony, which will associate
the GPS "time" message with the next PPS.

The ublox does not provide a low-latency time message that is not
affected by this. They all come after the navigation solution has
been computed. We therefore have to remove the ambiguity. This
suite of commits does so by allowing to increase the PPS period
to something much bigger than 1s (say, 5s)

The difficulty is that once the period is changed, we have to do
the necessary work to match the time info with the pulse. This is
done by listening to UBX-TIM-TP, which is a message that gives the
time of the next pulse. When the corresponding UBX-NAV-TIMEUTC
arrives, we know it is a time message related to a pulse and
send it to chrony.

The only remaining assumption is that the time-of-week part of
the TIMEUTC message is the same than the one from the TP message.
This is very unclear, as TIMEUTC's documentation lists it as "GPS
time of week of the navigation epoch" while TIM-TP says
"Time pulse time of week according to time base". I could verify
that they do match regardless of the configured time base (GPS or
UTC). Regardless, I have set that time base to GPS to be sure.