fix: change how the timing information is read in `chrony`
The current strategy was to read a timing information message, and then
wait for the corresponding PPS. The bug here is that we start reading
just after the last PPS, meaning we might be reading the next pulse
information that was received *before* the PPS (because hardware
buffering)

The last commit attempted to solve this particular issue. Which did not
solve the more general issue that if the pps/ublox synchronization
were to have a one or more period errors, we would not be able to
know and detect it.

This commit changes the strategy to poll the timing message, and use
the last timing message received whenever the PPS arrives. This is
not perfect either, but I believe chrony would "solve" outliers in
this case.