fix: test against PPS_CANWAIT in pps caps instead of params
The spec says params but it is readonly (which is ... weird),
but it seems that it should really be checked in the caps. It
does make more sense, and it is what gpsd does.

The params PPS_CANWAIT is cleared on boot on BBBs, but appears
magically once chrony ran at least once. My guess is that it
appears after the type of assertions has been configured.